init: optimized ffmpeg to only perform upscale by streaming
The existing algorithm is employing block algorithm that consumes
a lot of resources (as in storage requiring at least 3x for a video).
This is not desirable for large video file. Hence, we have to
optimize the algorithm to use streaming method that uses 1 sample
file while maintaining 1 finalized output file and 1 set of output
frames.

Also, we have to add the ability to recover from last run since video
file is usually massive.

Hence, let's do this.

This patch optimizes ffmpeg to only perform upscale 1 frame at a time
via streaming approach and adds a recovery method for longer and larger
video upscaling.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>