init/unix.sh: improved ffmpeg to only process frame
At the moment, ffmpeg manually breaks the video into components
at least twice: 1 for video and 1 for audio. This is time consuimg.
In fact, ffmpeg provides a method to rescale the original video
and allow the upscaled frame to passthrough it without needing to
deal with the audio component at all. Hence, this can save a lot
of time for large video. I should refactor it accordingly.

This patch improves ffmpeg to only process frame in init/unix.sh.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>