root: refactored polygot scripts for policy and security elevations
Appearently, after a preliminary testing, there was a security block
from the Windows side of things stating that the PowerShell was
unsigned. To overcome this when user downloded the repository locally,
we have to append '-ExecutionPolicy Bypass' argument into powershell
executables. This is better than instructing user to bring the entire
security system down. Hence, let's do this.

This patch refactors polygot scripts for policy and security elevations
in root/ directory.

Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>