root: initialized zoralab's monteur for automated CI packaging
As the git repository grew big due to binary blobs per patch, it
is not recommended to continue asking end-users to git clone the
repository. Hence, we seek zoralab's monteur's CI system to perform
automated packaging for this repository.

This patch initializes repository to use zoralba's monteur for
automated CI packaging in root/ repository.

Site-note: added Shuralyov, Jean because of the piggybacking for
GitHub pages. =.="

Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: Joly0 <13993216+Joly0@users.noreply.github.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>