README.md: make it a little more friendly for visitors
The marketing pitch is a bit too unfriendly for visitors. The writing
is targeting internal staffs rather than new people. Hence, we need
to make it a little more friendly so that people will use the project.

This patch makes the marketing pitch a little more friendly for visitor
in README.md.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <124406765+corygalyna@users.noreply.github.com>
Co-authored-by: Joly0 <13993216+Joly0@users.noreply.github.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>