init: fixed output path determination error
There was an error determining the output path where if the given
path is only a filepath (without directory), the program will try
to use the original input filename as a directory and cause the
upscaling process to be unreasonably failed. Hence, we need to fix
this bug.

This patch fixes the output path determination error in init/
directory.

Co-authored-by: Shuralyov, Jean <jean.shuralyov@proton.me>
Co-authored-by: Galyna, Cory <cory.galyna@gmail.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>