start.cmd: fixed parameters --input bug
It appears that from the ploygot script, the --input parameter
name is conflicting with the PowerShell "input" name and "path",
according to Joly0. Hence, this caused an inconsistency from
the end-user side where the provided "input" or "path" is not
the correct one as reported by Geoff McQueen.

This patch fixes the parameters --input naming convention bug
as stated above in the start.cmd polygot script.

Reported-by: Geoff McQueen [https://github.com/geoff-m]
Signed-off-by: Joly0 <13993216+Joly0@users.noreply.github.com>
Accepted-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Reworded-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Reviewed-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>