init: implemented parallel frame processing for windows.ps1
Since the Unix counterpart has parallel executions for video
processing, it's time to implement for windows.sp1. This only
affects the older implementations so that the project can get
going while having the new implementations unaffected.

Joly0 is the sole author of this patch. To awards his GitHub
contribution and due to GitHub's strict requirement, this patch
is written as "Co-authored" to make sure Joly0 gains the necessary
credit for his profile badge.

This patch implements parallel frame processing for windows.ps1
in init/ directory.

Co-authored-by: Joly0 <13993216+Joly0@users.noreply.github.com>
Co-authored-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: Joly0 <13993216+Joly0@users.noreply.github.com>
Accepted-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Reworded-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Reviewed-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>
Signed-off-by: (Holloway) Chew, Kean Ho <hollowaykeanho@gmail.com>